db: key converted to bigint (effort to decrease conflicts) and value converted to blob (extend the range of stored values)

+++ assets/postgres.cache.sql
@@ -1,5 +1,5 @@
 CREATE TABLE cache (
-	key INTEGER NOT NULL,
+	key BIGINT NOT NULL,
 	value TEXT NOT NULL,
 	PRIMARY KEY (key)
 );
